requirements.txt
Django==2.0.4
django-crispy-forms==1.7.2
django-localflavor==2.0
pytz==2018.4
